Initial commit
Added a mutex classes that uses drivers to provide differing
implementations for creating mutexes. The two provided are mutexes
created using semaphores and flock, with flock being the default, due
to restrictions on PHP semaphores.

Also added is a new session driver for concurrent sessions. This is a
cookie/database driver that supports value-, session-, and
application-level locking, along with session forwarding.